requirements.txt
pyTelegramBotAPI>=4.14.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sounddevice>=0.4.6
numpy>=1.24
